Moving files out of the Styling folder.
We should try to keep everything in one place. Make changes to my stuff
if you want to Guvenc- that's why we have version control, so we can
undo the changes if we want to go back, and we can merge changes if we
both make changes to the same file.